refactor: Remove old functions
