+++ packages/dc_solver_pkg/tests/export/test_asset_topology_to_dgs.py
@@ -28,17 +28,16 @@ from toop_engine_dc_solver.export.dgs_v7_definitions import (
     DGS_GENERAL_SHEET_CONTENT_FID_CIM,
     DgsElmCoupSchema,
 )
-from toop_engine_dc_solver.export.export import (
-    SwitchUpdateSchema,
+from toop_engine_dc_solver.postprocess.apply_asset_topo_powsybl import (
     get_asset_bay_grid_model_id_list,
     get_asset_switch_states_from_station,
     get_busbar_lookup,
     get_changing_switches_from_topology,
     get_coupler_states_from_busbar_couplers,
     get_diff_switch_states,
-    get_switch_update_schema_from_topology,
 )
 from toop_engine_interfaces.asset_topology import BusbarCoupler
+from toop_engine_interfaces.switch_update_schema import SwitchUpdateSchema
 
 
 def test_get_dgs_general_schema():
@@ -221,9 +220,14 @@ def test_get_busbar_lookup(basic_node_breaker_topology):
     assert busbar_lookup == expected
 
 
-def test_get_switch_update_schema_from_topology(basic_node_breaker_topology):
-    topology = basic_node_breaker_topology
-    switch_update_schema = get_switch_update_schema_from_topology(topology)
+def test_station_helpers_build_switch_update_schema(basic_node_breaker_topology):
+    station = basic_node_breaker_topology.stations[0]
+    coupler_df = get_coupler_states_from_busbar_couplers(station.couplers)
+    switch_reassignment_df, switch_disconnection_df = get_asset_switch_states_from_station(station)
+    switch_update_schema = cast(
+        pat.DataFrame[SwitchUpdateSchema],
+        pd.concat([coupler_df, switch_reassignment_df, switch_disconnection_df], ignore_index=True),
+    )
     expected = pd.DataFrame(
         [
             {"grid_model_id": "VL4_BREAKER", "open": True},
@@ -239,8 +243,13 @@ def test_get_switch_update_schema_from_topology(basic_node_breaker_topology):
 
 def test_get_diff_switch_states(basic_node_breaker_grid_v1, basic_node_breaker_topology):
     net = basic_node_breaker_grid_v1
-    topology = basic_node_breaker_topology
-    switch_update_schema = get_switch_update_schema_from_topology(topology)
+    station = basic_node_breaker_topology.stations[0]
+    coupler_df = get_coupler_states_from_busbar_couplers(station.couplers)
+    switch_reassignment_df, switch_disconnection_df = get_asset_switch_states_from_station(station)
+    switch_update_schema = cast(
+        pat.DataFrame[SwitchUpdateSchema],
+        pd.concat([coupler_df, switch_reassignment_df, switch_disconnection_df], ignore_index=True),
+    )
     diff_switch_states = get_diff_switch_states(network=net, switch_df=switch_update_schema)
     SwitchUpdateSchema.validate(diff_switch_states)
     expected = [
@@ -267,9 +276,14 @@ def test_get_changing_switches_from_topology(basic_node_breaker_grid_v1, basic_n
 
 
 def test_switch_update_schema_to_dgs(basic_node_breaker_grid_v1, basic_node_breaker_topology):
-    net = basic_node_breaker_grid_v1
-    realized_topology = basic_node_breaker_topology
-    switch_update_schema = get_switch_update_schema_from_topology(realized_topology)
+    _net = basic_node_breaker_grid_v1
+    station = basic_node_breaker_topology.stations[0]
+    coupler_df = get_coupler_states_from_busbar_couplers(station.couplers)
+    switch_reassignment_df, switch_disconnection_df = get_asset_switch_states_from_station(station)
+    switch_update_schema = cast(
+        pat.DataFrame[SwitchUpdateSchema],
+        pd.concat([coupler_df, switch_reassignment_df, switch_disconnection_df], ignore_index=True),
+    )
     foreign_ids = deepcopy(switch_update_schema)
     foreign_ids["foreign_id"] = foreign_ids["grid_model_id"] + "_foreign_id"
     foreign_ids.drop(columns=["open"], inplace=True)
