+++ packages/dc_solver_pkg/tests/export/test_disconnection_switch_updates.py
@@ -14,7 +14,8 @@ from toop_engine_dc_solver.export.disconnection_switch_updates import (
     get_disconnected_asset_ids,
     get_station_ids_affected_by_disconnections,
 )
-from toop_engine_dc_solver.export.export import get_changing_switches_from_actions, get_changing_switches_from_topology
+from toop_engine_dc_solver.export.export import get_changing_switches_from_actions
+from toop_engine_dc_solver.postprocess.apply_asset_topo_powsybl import get_changing_switches_from_topology
 from toop_engine_interfaces.nminus1_definition import GridElement
 from toop_engine_interfaces.switch_update_schema import SwitchUpdateSchema
